feat: use OpenTelemetry CICD Pipeline Attributes (#21)
> CICD Pipeline Attributes describes attributes specific to pipelines
within a Continuous Integration and Continuous Deployment (CI/CD)
system. A
[pipeline](https://en.wikipedia.org/wiki/Pipeline_%28computing%29) in
this case is a series of steps that are performed in order to deliver a
new version of software.
> This aligns with the
[Britannica](https://www.britannica.com/dictionary/pipeline) definition
of a pipeline where a pipeline is the system for developing and
producing something. In the context of CI/CD, a pipeline produces or
delivers software.

https://opentelemetry.io/docs/specs/semconv/attributes-registry/cicd/

Co-authored-by: Marcel Birkner <marcel.birkner@dash0.com>