feature: Support http as a trigger for `isHttpEndpoint` (#23)
This would extend `isHttpEndpoint` to allow `http` in addition to
`https`. This is particularly useful when the OTEL Collector resides
within the same VPC as self-hosted GitHub runners or ARC runners.

As such, traces can be sent directly to the OTEL Collector without
exposing it to the public internet. This also avoids unnecessary network
overhead as the traces do not need to be sent to the public internet to
then be sent back to the internal network if the GitHub runners,
self-hosted, are within the same VPC.