feat(trace): trickle down attributes from workflow to job and step
We need additional attributes to be added to the step and job spans to
make smarter filters in the SigNoz dashboards. As such, we can pass the
parent span attributes down to the child spans and remove the individual
name attributes that were previously added to the step and job spans since
they are now redundant.